requirements.txt
flask==3.1.2
flask-cors==4.0.0
gunicorn==21.2.0
pillow==11.3.0
numpy
matplotlib
gdown        
tensorflow 
